Restructuracion del proyecto uasndo la estructura de flutter-skeleton.
