+++ lib/design/app_colors.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+mixin AppColors {
+
+  static const Color custom_red = Color(0xFFD52F23);
+
+  static const Color custom_ambar = Color(0xFFAAAA11);
+
+}
