+++ lib/screens/pelicula_detalle.dart
