+++ lib/search/search_delegate.dart
@@ -1,9 +1,7 @@
 
-
-
 import 'package:flutter/material.dart';
 import 'package:peliculas_bitbox/models/pelicula_modelo.dart';
-import 'package:peliculas_bitbox/providers/peliculas_providers.dart';
+import 'package:peliculas_bitbox/repository/peliculas_providers.dart';
 
 class DataSearch extends SearchDelegate {
